modify ActionManager to be used as a  MixIn class
moved all related functionnality in the action_manager module
